Change information shown on candidate list page
The most important thing to know isn't *have we processed people's
applications* (the committee doesn't use 'potential' to indicate that
people have been reviewed, or 'not recommended' really ever, so people
may never leave 'applicant'). The important thing to know is *whether
we have facilitated involvement*.

So we're categorizing people by how many appointments (recommended or
actual) they have, so that we can tell if we're missing or overloading
people.